Overhaul of all the documentation strings. (#30)
Cleaning up docstrings, fighting with sphinx on a number of issues related to
type aliases & links to external packages. type hints seem to be extremely problematic
for descriptions. We have done our best to make sure most objects (classes,
functions) are linked both within the package and to external packages.

---------

Co-authored-by: jkanche <jayaram.kancherla@gmail.com>