Added dry-run capability to the analyze overlord function. (#32)
We add a script to parse the live _analyze function, which creates a new
function that generates the dry run code. This avoids the need for manual
synchronization between the live and dry code (or even worse, eval()).

We reorganized the code to move some of the work out of the analyze function so
that the dry run doesn't spew out too much complex logic. Specifically, the
custom threshold overrides are now moved into suggest_rna_qc_filters, and the
parallelized UMAP/t-SNE is now in its own function.

Updated tests and documentation correspondingly.

Co-authored-by: LTLA <infinite.monkeys.with.keyboards@gmail.com>