Shift the sanitization arguments to center_size_factors. (#39)
Now log_norm_counts is just a thin wrapper around center_size_factors. This
simplifies the choice between DelayedArray and TatamiNumericPointer outputs, as
the same set of size factors are always used between the two choices.

This involves an update to the libscran version to get the new sanitization code.