Improve execution flow for running neighborhood suite operations (#79).
Modifying the execution of neighborhood suite operations (computing t-SNE, UMAP, and SNN graph). When the package is run on macOS, we switch to "fork" to circumvent the need for `if __name__ == '__main__':` in user scripts. We completely avoid multiprocessing and execute all operations sequentially if the number of available threads is 1.