Overhaul to use the new libscran libraries via pybind11 (#83)
We no longer use our cpptypes wrapper, instead using pybind11 for convenience.
We also switch over to CMake for a more convenient compilation.

The functions in the new package are now aligned with the scrapper BioC package
for consistency. This entails the replacement of some functions (e.g., 
grouped_size_factors by compute_clrm1_factors), the renaming of others,
(e.g., score_feature_set is now score_gene_set), and some new functions
(e.g., cluster_kmeans, aggregate_across_genes).

All functions have been simplified to remove the Options classes. Instead,
people can just parametrize each step via regular function arguments.

For functions with complicated outputs, the return value is now a specialized
*Results dataclass, with an method to optionally convert it to a BiocPy class.
This is easier to document and avoids a hard dependency on BiocPy classes while
preserving interoperability when the BiocPy classes are needed.

We build the igraph static library and link to it ourselves, and use this to
enable clustering in cluster_graph(). This avoids unanticipated differences
in results due to changes in the igraph dependency. Nonetheless, our graph
objects can still be converted to igraph.Graph for interoperability.

We removed the metaprogramming for dry runs of the analyze() workflow; users
can directly inspect the source code if they are interested.

All neighbor search code has been removed in favor of a dependency on the
knncolle package, which avoids the need to recompile search code for scranpy.

The various CI workflows have been updated with the new build requirements,
and to remove support for the EOL'd Python 3.8.

Co-authored-by: Jayaram Kancherla <jayaram.kancherla@gmail.com>